requirements.txt
torchmetrics==1.3.1
tabulate==0.9.0
tqdm==4.65.0
numpy==1.24.3
matplotlib==3.7.0
# torch==1.12.1 # Install following the instructions in https://pytorch.org/